+++ src/java/cz/cuni/mff/d3s/spl/data/BenchmarkRunUtils.java
@@ -93,7 +93,7 @@ public class BenchmarkRunUtils  {
 		synchronized (runs) {
 			for (BenchmarkRun r : runs) {
 				synchronized (r) {
-					for (long l : r.getSamples()) {
+					for (double l : r.getSamples()) {
 						builder.addSamples(l);
 					}
 				}
